+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 ==============
 
+
 1.0.0-alpha28.pre
 -----------------
  * Updated yandex kassa